# swagger.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: My blogAPI
  description: This is a very simple and minimalistic blog api for posting articles and get comment from users
servers:
- url: http://localhost:4000/api/v1
  description: dev serve
...
paths:
...
  /articles/{id}:
    get:
      tags:
      - Article
      description: Fetch a single article by ID
      parameters:
      - name: id
        required: true
        in: path
        description: article ID
        schema:
          type: string
          examples:
          - 125455
      - name: expand
        in: query
        description: to expand related data
        schema:
          type: string
          enum:
          - author
          - comment
          examples:
          - author
      responses:
        "200":
          description: successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                    - type: object
                      required:
                      - title
                      - body
                      properties:
                        title:
                          type: string
                          examples:
                          - how to start passive income
                        body:
                          type: string
                          format: text
                          examples:
                          - this is the body of article
                        cover:
                          type: string
                          format: url
                          examples:
                          - https://s3.aws.com/myblogapi/cover/artcle1.png
                        status:
                          type: string
                          enum:
                          - draft
                          - published
                          default: draft
                    - type: object
                      properties:
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                    - type: object
                      properties:
                        author:
                          type: object
                          required:
                          - id
                          - name
                          properties:
                            id:
                              type: string
                              examples:
                              - '6585xe74seq454ax26975'
                            name:
                              type: string
                              examples:
                              - Raybrin Max
                  comments:
                          type: array
                          items:
                            allOf:
                            - type: object
                              required:
                              - title
                              - body
                              properties:
                                id:
                                  type: string
                                  examples:
                                  - '6585xe74seq454ax26975'
                                postId:
                                  type: string
                                  examples:
                                  - '6585xe74seq454ax26975'
                                authorId:
                                  type: string
                                  examples:
                                  - '6585xe74seq454ax26975'
                                body:
                                  type: string
                                  format: text
                                  examples:
                                  - this is the body of comments
                                status:
                                  type: string
                                  enum:
                                  - public
                                  - private
                                  default: public
                            - type: object
                              properties:
                                createdAt:
                                  type: string
                                  format: date-time
                                updatedAt:
                                  type: string
                                  format: date-time  
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        examples:
                        - /articles/1
                      author:
                        type: string
                        examples:
                        - /articles/1/author
                      comment:
                        type: string
                        examples:
                        - /articles/1/comments
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 400
                  error:
                    type: string
                    examples:
                    - Bad request
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          examples:
                          - email
                        message:
                          type: string
                          examples:
                          - invalid email
        "404":
          description: Resource Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 404
                  error:
                    type: string
                    examples:
                    - Not Found
                  message:
                    type: string
                    examples:
                    - Request resource not found
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 500
                  error:
                    type: string
                    examples:
                    - Internal Server Error
                  message:
                    type: string
                    examples:
                    - We are sorry for the inconvenience. Please try again
    put:
      security:
      - bearerauth: []
      tags:
      - Article
      description: create or update entire article
      parameters:
      - name: id
        required: true
        in: path
        description: authorId
        schema:
          type: string
          examples:
          - 125455
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
              - title
              - body
              properties:
                title:
                  type: string
                  examples:
                  - how to start passive income
                body:
                  type: string
                  format: text
                  examples:
                  - this is the body of article
                cover:
                  type: string
                  format: url
                  examples:
                  - https://s3.aws.com/myblogapi/cover/artcle1.png
                status:
                  type: string
                  enum:
                  - draft
                  - published
                  default: draft
      responses:
        "200":
          description: successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 200
                  message:
                    type: string
                    examples:
                    -  successfully updated
                  data:
                    allOf:
                    - type: object
                      required:
                      - title
                      - body
                      properties:
                        title:
                          type: string
                          examples:
                          - how to start passive income
                        body:
                          type: string
                          format: text
                          examples:
                          - this is the body of article
                        cover:
                          type: string
                          format: url
                          examples:
                          - https://s3.aws.com/myblogapi/cover/artcle1.png
                        status:
                          type: string
                          enum:
                          - draft
                          - published
                          default: draft
                    - type: object
                      properties:
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        examples:
                        - /articles/1
        "201":
          description: successfully Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 201
                  message:
                    type: string
                    examples:
                    - successfully created
                  data:
                    allOf:
                    - type: object
                      required:
                      - title
                      - body
                      properties:
                        title:
                          type: string
                          examples:
                          - how to start passive income
                        body:
                          type: string
                          format: text
                          examples:
                          - this is the body of article
                        cover:
                          type: string
                          format: url
                          examples:
                          - https://s3.aws.com/myblogapi/cover/artcle1.png
                        status:
                          type: string
                          enum:
                          - draft
                          - published
                          default: draft
                    - type: object
                      properties:
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        examples:
                        - /articles/1
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 400
                  error:
                    type: string
                    examples:
                    - Bad request
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          examples:
                          - email
                        message:
                          type: string
                          examples:
                          - invalid email
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 401
                  error:
                    type: string
                    examples:
                    - Unauthorized request
                  message:
                    type: string
                    examples:
                    - You don't have the permission
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 500
                  error:
                    type: string
                    examples:
                    - Internal Server Error
                  message:
                    type: string
                    examples:
                    - We are sorry for the inconvenience. Please try again
    patch:
      security:
      - bearerauth: []
      tags:
      - Article
      description: update existing properties of an article
      parameters:
      - name: id
        required: true
        in: path
        description: authorId
        schema:
          type: string
          examples:
          - 125455
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  examples:
                  - how to start passive income
                body:
                  type: string
                  format: text
                  examples:
                  - this is the body of article
                cover:
                  type: string
                  format: url
                  examples:
                  - https://s3.aws.com/myblogapi/cover/artcle1.png
                status:
                  type: string
                  enum:
                  - draft
                  - published
                  default: draft
      responses:
        "200":
          description: successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 200
                  message:
                    type: string
                    examples:
                    - Article successfully updated
                  data:
                    allOf:
                    - type: object
                      properties:
                        title:
                          type: string
                          examples:
                          - how to start passive income
                        body:
                          type: string
                          format: text
                          examples:
                          - this is the body of article
                        cover:
                          type: string
                          format: url
                          examples:
                          - https://s3.aws.com/myblogapi/cover/artcle1.png
                        status:
                          type: string
                          enum:
                          - draft
                          - published
                          default: draft
                    - type: object
                      properties:
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        examples:
                        - /articles/1
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 400
                  error:
                    type: string
                    examples:
                    - Bad request
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          examples:
                          - email
                        message:
                          type: string
                          examples:
                          - invalid email
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 401
                  error:
                    type: string
                    examples:
                    - Unauthorized request
                  message:
                    type: string
                    examples:
                    - You don't have the permission
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    examples:
                    - 500
                  error:
                    type: string
                    examples:
                    - Internal Server Error
                  message:
                    type: string
                    examples:
                    - We are sorry for the inconvenience. Please try again
    delete:
      security: 
        - bearerauth: []
      tags: 
        - Article
      description: delete an article
      parameters: 
        - $ref: '#/components/parameters/idPathParam'
      responses: 
        '204':
          description: Article deleted successfully
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    bearerauth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    "401":
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                examples:
                - 401
              error:
                type: string
                examples:
                - Unauthorized request
              message:
                type: string
                examples:
                - You don't have the permission
    "404":
      description: Resource Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                examples:
                - 404
              error:
                type: string
                examples:
                - Not Found
              message:
                type: string
                examples:
                - Request resource not found
    "500":
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                examples:
                - 500
              error:
                type: string
                examples:
                - Internal Server Error
              message:
                type: string
                examples:
                - We are sorry for the inconvenience. Please try again
  parameters:
...
    idPathParam:
      name: id
      required: true
      in: path
      description: Id
      schema:
        type: string
        examples:
        - 125455